user interface/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
SQLAlchemy==2.0.37
PyMySQL==1.1.1
python-multipart==0.0.20
python-dotenv==1.1.0
boto3==1.38.21
botocore==1.38.21
pydantic==2.10.4
pydantic-settings==2.9.1

